requirements.txt
fastapi
uvicorn[standard]
python-multipart
faiss-cpu
PyMuPDF
tiktoken
pandas
requests
aiohttp
